need this tested on arm device

+++ openhcl_kernel.nix
@@ -1,8 +1,13 @@
-{ system, stdenv, fetchzip, is_dev ? false, is_cvm ? false, }:
+{ system, stdenv, fetchzip, targetArch ? null, is_dev ? false, is_cvm ? false }:
 
 let
   version = if is_dev then "6.12.44.1" else "6.12.44.1";
-  arch = if system == "aarch64-linux" then "arm64" else "x64";
+  # Allow explicit override of architecture, otherwise derive from host system
+  # Note: targetArch uses "x86_64"/"aarch64", but URLs use "x64"/"arm64"
+  arch = if targetArch == "x86_64" then "x64"
+         else if targetArch == "aarch64" then "arm64"
+         else if system == "aarch64-linux" then "arm64"
+         else "x64";
   branch = if is_dev then "hcl-dev" else "hcl-main";
   build_type = if is_cvm then "cvm" else "std";
   # See https://github.com/microsoft/OHCL-Linux-Kernel/releases
@@ -35,7 +40,7 @@ let
   hash = hashes.${branch}.${build_type}.${arch};
 
 in stdenv.mkDerivation {
-  pname = "openhcl-kernel";
+  pname = "openhcl-kernel-${arch}";
   inherit version;
   src = fetchzip {
     inherit url;
@@ -46,7 +51,13 @@ in stdenv.mkDerivation {
   installPhase = ''
     runHook preInstall
     mkdir -p $out/modules
-    cp vmlinux* $out/
+    # x64 uses vmlinux, arm64 uses Image
+    if [ -f vmlinux ]; then
+      cp vmlinux* $out/
+    fi
+    if [ -f Image ]; then
+      cp Image $out/
+    fi
     cp -r modules/* $out/modules/
     cp kernel_build_metadata.json $out/
     runHook postInstall
